services/scheduler_service/requirements.txt
psycopg2-binary==2.9.9
pika==1.3.2
python-dateutil==2.8.2
pytz==2024.1
requests==2.31.0
python-dotenv==1.0.0
